requirements.txt
langchain==0.3.27
langchain-core==0.3.79
langchain-openai==0.3.35
openai==2.3.0
requests==2.32.5
python-dotenv==1.1.1
pydantic==2.12.0
pydantic_core==2.41.1
typing-extensions==4.15.0
